tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/send/012.html

<!doctype html>
<title>WebSockets: sending null</title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>

async_test(function(t){
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo');
  ws.onopen = t.step_func(function(e) {
    ws.send(null);
  });
  ws.onmessage = t.step_func(function(e) {
    assert_equals(e.data, 'null');
    ws.onclose = t.step_func(function(e) {
      ws.onclose = t.unreached_func();
      t.step_timeout(() => t.done(), 50);
    })
    ws.close();
  });
  ws.onclose = ws.onerror = t.unreached_func();
});
</script>
